+++ src/screens/transaction_details/components/index.ts
@@ -1,31 +1,31 @@
 import Overview from './overview';
 import Messages from './messages';
-import Delegate from './delegate';
-import Unknown from './unknown';
-import Redelegate from './redelegate';
-import Undelegate from './undelegate';
-import CreateValidator from './create_validator';
-import EditValidator from './edit_validator';
-import Send from './send';
-import Multisend from './multisend';
-import VerifyInvariant from './verify_invariant';
-import Unjail from './unjail';
-import Fund from './fund';
-import SetWithdrawalAddress from './set_withdrawal_address';
-import WithdrawReward from './withdraw_reward';
-import DepositProposal from './deposit_proposal';
-import Vote from './vote';
-import SubmitProposal from './submit_proposal';
-import WithdrawCommission from './withdraw_commission';
-import SaveProfile from './save_profile';
-import DeleteProfile from './delete_profile';
-import CreateRelationship from './create_relationship';
-import DtagTransferRequest from './dtag_transfer_request';
-import DtagAcceptTransfer from './dtag_accept_transfer';
-import DtagCancelTransfer from './dtag_cancel_transfer';
-import DtagRefuseTransfer from './dtag_refuse_transfer';
-import BlockUser from './block_user';
-import UnBlockUser from './unblock_user';
+import Delegate from './msg/staking/delegate';
+import Unknown from './msg/unknown';
+import Redelegate from './msg/staking/redelegate';
+import Undelegate from './msg/staking/undelegate';
+import CreateValidator from './msg/staking/create_validator';
+import EditValidator from './msg/staking/edit_validator';
+import Send from './msg/bank/send';
+import Multisend from './msg/bank/multisend';
+import VerifyInvariant from './msg/crisis/verify_invariant';
+import Unjail from './msg/slashing/unjail';
+import Fund from './msg/distribution/fund';
+import SetWithdrawalAddress from './msg/distribution/set_withdrawal_address';
+import WithdrawReward from './msg/distribution/withdraw_reward';
+import DepositProposal from './msg/governance/deposit_proposal';
+import Vote from './msg/governance/vote';
+import SubmitProposal from './msg/governance/submit_proposal';
+import WithdrawCommission from './msg/distribution/withdraw_commission';
+import SaveProfile from './msg/profiles/save_profile';
+import DeleteProfile from './msg/profiles/delete_profile';
+import CreateRelationship from './msg/profiles/create_relationship';
+import DtagTransferRequest from './msg/profiles/dtag_transfer_request';
+import DtagAcceptTransfer from './msg/profiles/dtag_accept_transfer';
+import DtagCancelTransfer from './msg/profiles/dtag_cancel_transfer';
+import DtagRefuseTransfer from './msg/profiles/dtag_refuse_transfer';
+import BlockUser from './msg/profiles/block_user';
+import UnBlockUser from './msg/profiles/unblock_user';
 
 export {
   Overview,
