+++ src/screens/transaction_details/components/msg/bank/multisend/index.tsx
