+++ src/screens/transaction_details/components/msg/distribution/fund/index.tsx
