+++ src/screens/transaction_details/components/msg/distribution/set_withdrawal_address/index.tsx
