+++ src/screens/transaction_details/components/msg/distribution/withdraw_commission/index.tsx
