+++ src/screens/transaction_details/components/msg/distribution/withdraw_reward/index.tsx
