+++ src/screens/transaction_details/components/msg/governance/deposit_proposal/index.tsx
