+++ src/screens/transaction_details/components/msg/governance/submit_proposal/index.tsx
