+++ src/screens/transaction_details/components/msg/profiles/dtag_accept_transfer/index.tsx
