+++ src/screens/transaction_details/components/msg/profiles/dtag_cancel_transfer/index.tsx
