+++ src/screens/transaction_details/components/msg/profiles/dtag_refuse_transfer/index.tsx
