+++ src/screens/transaction_details/components/msg/staking/create_validator/index.tsx
