+++ src/screens/transaction_details/components/msg/staking/delegate/index.tsx
