+++ src/screens/transaction_details/components/msg/staking/edit_validator/index.tsx
