+++ src/screens/transaction_details/components/msg/staking/redelegate/index.tsx
